refactor: move ColoringModes enum to common, so that importing OSMDOptions doesn't cause big import chain
